[UPDATE] 검색 페이지를 삭제하고, 필터페이지로 합치고 Search Page로 이름 변경, SaerchPage와 SearchResultPage 연동
