+++ modak/lib/main.dart
@@ -8,7 +8,6 @@ import 'package:logger/logger.dart';
 import 'package:modak/bloc/CampingAPIBloc.dart';
 import 'package:modak/bloc/ModakBloc.dart';
 import 'package:modak/bloc/UserBloc.dart';
-import 'package:modak/page/CampingSearchPage.dart';
 import 'package:modak/page/CampingSearchResultPage.dart';
 import 'package:modak/page/CapingDetailPage.dart';
 import 'package:modak/page/CreateMatchingPage.dart';
@@ -73,16 +72,6 @@ class MyApp extends StatelessWidget {
           fontFamily: 'BMDOHYEON_ttf',
         ),
         initialRoute: '/splash',
-        onGenerateRoute: (settings) {
-          var routeName = settings.name;
-          switch(routeName) {
-            case '/camping_search_result':
-              Object? param = settings.arguments;
-
-              String stringParam = param.toString();
-              return MaterialPageRoute(builder: (_) => CampingSearchResultPage(searchName: stringParam));
-          }
-        },
         routes: {
           '/splash': (context) => SplashPage(),
           '/dashboard': (context) => DashBoardPage(),
@@ -92,8 +81,8 @@ class MyApp extends StatelessWidget {
           '/payment': (context) => PaymentPage(),
           '/camping_detail': (context) => CampingDetailPage(),
           '/create_matching': (context) => CreateMatchingPage(),
-          '/camping_search': (context) => CampingSearchPage(),
           '/goods': (context) => GoodsPage(),
+          '/camping_search_result': (context) => CampingSearchResultPage(),
         },
       ),
     );
