+++ modak/lib/page/CampingSearchPage.dart
@@ -1,24 +1,66 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:modak/bloc/CampingAPIBloc.dart';
+import 'package:modak/bloc/CampingAPIEvent.dart';
+import 'package:modak/bloc/CampingAPIState.dart';
 
 void main() {
   runApp(MaterialApp(
     home: Scaffold(
       body: Center(
-        child: CampingSearchPage(),
+        child: CampingFilterPage(),
       ),
     ),
   ));
 }
 
-class CampingSearchPage extends StatefulWidget {
+class CampingFilterPage extends StatefulWidget {
   final TextEditingController _searchController = TextEditingController();
 
   @override
-  CampingSearchPageState createState() => CampingSearchPageState();
+  CampingFilterPageState createState() => CampingFilterPageState();
 }
 
-class CampingSearchPageState extends State<CampingSearchPage> {
+class CampingFilterPageState extends State<CampingFilterPage> {
+  Widget _filterButton(String title) {
+    return Container(
+      margin: const EdgeInsets.all(8.0),
+      child: InkWell(
+        onTap: () {},
+        child: Ink(
+          padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 20),
+          child: Text(
+            title,
+            maxLines: 1,
+            style: TextStyle(color: Colors.black),
+          ),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.all(
+              Radius.circular(2.0),
+            ),
+            boxShadow: [
+              BoxShadow(
+                color: Color(0x22000000),
+                offset: Offset(2.0, 2.0),
+                blurRadius: 2.0,
+                spreadRadius: 0.1,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  @override
+  void initState() {
+    BlocProvider.of<CampingAPIBloc>(context).add(
+      GetCampingsFilterDataEvent(),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
     final double _width = MediaQuery.of(context).size.width;
@@ -77,6 +119,136 @@ class CampingSearchPageState extends State<CampingSearchPage> {
                 ],
               ),
             ),
+          ),
+          SizedBox(
+            height: 48.0,
+          ),
+          Container(
+            child: Row(
+              children: [
+                SizedBox(
+                  width: 20.0,
+                ),
+                Text("지역별",
+                    style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20))
+              ],
+            ),
+          ),
+          SizedBox(
+            height: 24.0,
+          ),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 20.0),
+            child: BlocBuilder<CampingAPIBloc, CampingAPIState>(
+                builder: (_, state) {
+              if (state is Empty) {
+                return Container();
+              } else if (state is Loading) {
+                return Center(child: CircularProgressIndicator());
+              } else if (state is Error) {
+                return Text("Error: ");
+              } else if (state is Loaded) {
+                // return Text(state.dataOperationTypes.toString());
+                return Row(
+                  children: [
+                    Flexible(
+                        child: Wrap(
+                            children: List.generate(
+                                    7,
+                                    (i) =>
+                                        _filterButton(state.dataRegions![i])) +
+                                [_filterButton("+더보기")])),
+                  ],
+                );
+              }
+              return Container();
+            }),
+          ),
+          SizedBox(
+            height: 24.0,
+          ),
+          Container(
+            child: Row(
+              children: [
+                SizedBox(
+                  width: 20.0,
+                ),
+                Text("운영 형태",
+                    style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20))
+              ],
+            ),
+          ),
+          SizedBox(
+            height: 24.0,
+          ),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 20.0),
+            child: BlocBuilder<CampingAPIBloc, CampingAPIState>(
+                builder: (_, state) {
+              if (state is Empty) {
+                return Container();
+              } else if (state is Loading) {
+                return Center(child: CircularProgressIndicator());
+              } else if (state is Error) {
+                return Text("Error: ");
+              } else if (state is Loaded) {
+                // return Text(state.dataOperationTypes.toString());
+                return Row(
+                  children: [
+                    Flexible(
+                        child: Wrap(
+                            children: List.generate(
+                                state.dataOperationTypes!.length,
+                                (i) => _filterButton(
+                                    state.dataOperationTypes![i])))),
+                  ],
+                );
+              }
+              return Container();
+            }),
+          ),
+          SizedBox(
+            height: 24.0,
+          ),
+          Container(
+              child: Row(children: [
+            SizedBox(
+              width: 20.0,
+            ),
+            Text("입지 구분",
+                style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20))
+          ])),
+          SizedBox(
+            height: 24.0,
+          ),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 20.0),
+            child: BlocBuilder<CampingAPIBloc, CampingAPIState>(
+                builder: (_, state) {
+              if (state is Empty) {
+                return Container();
+              } else if (state is Loading) {
+                return Center(child: CircularProgressIndicator());
+              } else if (state is Error) {
+                return Text("Error: ");
+              } else if (state is Loaded) {
+                // return Text(state.dataOperationTypes.toString());
+                return Row(
+                  children: [
+                    Flexible(
+                        child: Wrap(
+                            children: List.generate(
+                                state.dataEnvironments!.length,
+                                (i) => _filterButton(
+                                    state.dataEnvironments![i])))),
+                  ],
+                );
+              }
+              return Container();
+            }),
+          ),
+          SizedBox(
+            height: 150.0,
           ),
         ],
       ),
